+++ test-scripts/single-anchor-test.sh
@@ -6,8 +6,8 @@ fi
 
 export ANCHOR_WALLET=~/.config/solana/id.json
 
-test_files=(placeAndMakeSwiftPerpBankrun.ts)
+test_files=(fuelSweep.ts)
 
 for test_file in ${test_files[@]}; do
   ts-mocha -t 300000 ./tests/${test_file}
-done 
+done